Update wx code to use general Wildcat coding standards.  Also subverted wx Toolbar and ToolbarButton to conform to general standards so that no code is impacted outside of platform specific code.  My skills (or lack there of) have resulted in a number of bugs that Dan can hopefully work out.
The trimmed surfaces are still not rendering properly.  Need to fix this.

+++ trunk/Source/Utility/texture_manager_win32.cpp
@@ -72,8 +72,8 @@ void WCTextureManager::LoadTexture(WSTexture *texObj) {
 	}
 	else {
 		// use pow(2, n) for texture width and height
-		for(texObj->_texture_width = 1; texObj->_texture_width < img.width; texObj->_texture_width *= 2){}
-		for(texObj->_texture_height = 1; texObj->_texture_height < img.height; texObj->_texture_height *= 2){}
+		for(texObj->_texture_width = 1; texObj->_texture_width < (GLint)img.width; texObj->_texture_width *= 2){}
+		for(texObj->_texture_height = 1; texObj->_texture_height < (GLint)img.height; texObj->_texture_height *= 2){}
 	}
 
 	//Get all of the raster data
